aspect ratio scaling
looks bad, but we're sticking with it since I don't normally make web games. Also there's this weird lag behind and I'm not sure what it is... Might just be because the game state is being drawn with lag cause firefox's profiler says it's running at a relatively average framerate... Hmmm. It appears consistent at least?